feat: unit test for controller of plafor module
